move global components into ubiquitout/global
